Refactored food details activity
- using ButterKnife injections
- unified naming
- change package
