+++ app/src/main/java/com/canteen/app/activity/food/details/FoodDetailsActivityAdmin.java
@@ -1,4 +1,4 @@
-package com.canteen.app.activity.foodDetails;
+package com.canteen.app.activity.food.details;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,26 +15,25 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.canteen.app.App;
-import com.canteen.app.service.order.OrderCartService;
 import com.canteen.app.R;
 import com.canteen.app.activity.cart.OrderCartActivity;
 import com.canteen.app.models.Food;
-import com.canteen.app.service.order.OrderItem;
 
 import java.util.Locale;
 
-public class FoodDetailsActivity extends AppCompatActivity implements PriceContainer {
+// TODO: move or remove?
+public class FoodDetailsActivityAdmin extends AppCompatActivity implements PriceContainer {
 
     private Food food;
     private TextView foodDescriptionTextView;
     private TextView foodPriceTextView;
     private RecyclerView foodAdditionsRecyclerView;
-    private Button addToCartButton;
+    private Button deleteFood;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_food_details);
+        setContentView(R.layout.activity_food_details_admin);
 
         this.retrieveFood();
         this.initView();
@@ -45,7 +44,7 @@ public class FoodDetailsActivity extends AppCompatActivity implements PriceConta
     public boolean onCreateOptionsMenu(Menu menu) {
         super.onCreateOptionsMenu(menu);
         MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.menu, menu);
+        inflater.inflate(R.menu.menu_food, menu);
         return true;
     }
 
@@ -61,10 +60,10 @@ public class FoodDetailsActivity extends AppCompatActivity implements PriceConta
     }
 
     private void initView() {
-        this.foodDescriptionTextView = findViewById(R.id.foodDescriptionTextView);
-        this.foodAdditionsRecyclerView = findViewById(R.id.foodAdditionsRecyclerView);
-        this.foodPriceTextView = findViewById(R.id.foodPriceTextView);
-        this.addToCartButton = findViewById(R.id.addToCartButton);
+        this.foodDescriptionTextView = findViewById(R.id.foodDescriptionTextViewAdmin);
+        this.foodAdditionsRecyclerView = findViewById(R.id.foodAdditionsRecyclerViewAdmin);
+        this.foodPriceTextView = findViewById(R.id.foodPriceTextViewButton);
+        this.deleteFood = findViewById(R.id.deleteFoodButton);
 
         ActionBar actionBar = getSupportActionBar();
         if (actionBar != null) {
@@ -78,13 +77,14 @@ public class FoodDetailsActivity extends AppCompatActivity implements PriceConta
         this.foodAdditionsRecyclerView.setAdapter(adapter);
 
         this.foodDescriptionTextView.setText(this.food.getDescription());
-        this.addToCartButton.setOnClickListener(v -> {
-            OrderCartService.getInstance().addItem(OrderItem.builder()
-                    .food(food)
-                    .additions(adapter.getSelected())
-                    .build());
-            Toast.makeText(App.getContext(), "Food added to cart", Toast.LENGTH_SHORT)
-                    .show();
+        this.deleteFood.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+                Toast.makeText(App.getContext(), "Food delete", Toast.LENGTH_SHORT)
+                        .show();
+                finish();
+            }
         });
     }
