+++ app/src/main/java/com/canteen/app/activity/food/details/PriceContainer.java
@@ -1,4 +1,4 @@
-package com.canteen.app.activity.foodDetails;
+package com.canteen.app.activity.food.details;
 
 interface PriceContainer {
     void updatePrice(double priceIncrease);
