+++ app/src/main/res/layout/list_item_order_item.xml
@@ -13,7 +13,7 @@
         android:layout_marginEnd="8dp"
         android:text="@string/example_food_name"
         android:textAppearance="@android:style/TextAppearance.Material.Large"
-        app:layout_constraintEnd_toStartOf="@+id/foodPriceTextView"
+        app:layout_constraintEnd_toStartOf="@+id/food_price_text_view"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -30,7 +30,7 @@
         app:layout_constraintVertical_bias="0.0" />
 
     <TextView
-        android:id="@+id/foodPriceTextView"
+        android:id="@+id/food_price_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginEnd="8dp"
